Updated storage_mgmt_net_to_compute.patch for OSP8
In OSP7 the patch would change the following:

 templates-net/puppet/compute-puppet.yaml
 templates-net/compute.yaml
 templates-net/environments/network-isolation.yaml
 templates-net/overcloud-resource-registry-puppet.yaml

In OSP8 the last two files do not need to be patched because the
compute role already has the StorageMgmtPort, thus the new patch file
does not include changes for them.

The first two files seem to be the same as in OSP7 and have been moved
to a new directory "kilo" (I don't yet know why we put this in
Liberty). The semantics of the original patch work fine there and I
manually the changes using the following process to create the new
storage_mgmt_net_to_compute.patch for OSP8.

 cp -r /usr/share/openstack-tripleo-heat-templates ~/templates-net
 cp -r /usr/share/openstack-tripleo-heat-templates ~/templates

 vi templates-net/kilo/compute.yaml
 vi templates-net/kilo/puppet/compute-puppet.yaml

 diff -Naur /home/stack/templates /home/stack/templates-net > storage_mgmt_net_to_compute.patch